e2e_project/requirements.txt
django-avatar>=8.0.1,<8.1
lorem-text>=3.0,<3.1
markdown>=3.10,<3.11
